move Message types from server to utils
