+++ lamarrs-utils/src/messages.rs
@@ -1,7 +1,7 @@
 // Collection of messages types to be send via websockets.
 
 use serde::{Deserialize, Serialize};
-use crate::enums::{RelativeLocation, Service};
+use crate::enums::{Color, RelativeLocation, Service};
 use arrayvec::ArrayString;
 use uuid::Uuid;
 
@@ -17,4 +17,16 @@ pub struct Suscribe{
 #[serde(rename_all = "camelCase")]
 pub struct Subtitle{
     pub subtitle: ArrayString<35>,
+}
+
+#[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
+pub struct SendSubtitle {
+    pub subtitle: Subtitle,
+    pub target_location: RelativeLocation,
+}
+
+#[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
+pub struct SendColor {
+    pub color: Color,
+    pub target_location: RelativeLocation,
 }
